pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
fastapi = "^0.112.0"
uvicorn = "0.23.2"
openai = "^1.71.0"
stripe = "^12.0.0"
langchain = "^0.3.23"
sqlmodel = "^0.0.24"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.9"
langchain-community = "^0.3.21"
jinja2 = "^3.1.6"
sqlalchemy = "^2.0.40"
langchain-openai = "^0.3.12"
stytch = "^13.1.0"
pydantic = {extras = ["email"], version = "^2.11.3"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"
pytest-mock = "^3.11.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
